apps/zero/frontend/AngularAppCurrent/src/app/shared/wml-components/wml-fields/wml-error/wml-error.component.spec.ts
// testing
import { ComponentFixture } from '@angular/core/testing';
import { configureTestingModuleForComponents, grabComponentInstance, mockTranslateService } from '@app/core/utility/test-utils';

// rxjs
import { Subject } from 'rxjs';

import { WmlErrorComponent } from './wml-error.component';


describe('WmlErrorComponent', () => {
  let cpnt: WmlErrorComponent;
  let fixture: ComponentFixture<WmlErrorComponent>;

  beforeEach(async () => {
    await configureTestingModuleForComponents(WmlErrorComponent,{mockTranslateService});
    ({fixture, cpnt} =  grabComponentInstance(WmlErrorComponent));
    fixture.detectChanges()
  })

  describe("init", () => {

    it("should create", () => {
      expect(cpnt).toBeTruthy()
    })  

    it("should have all values initalize properly", () => {
      expect(cpnt.myClass).toEqual('View')
    })

    it("should have all properties be the correct class instance", () => {
      expect(cpnt.ngUnsub).toBeInstanceOf(Subject<void>)
    })
  })

  describe("ngOnDestroy",()=>{

    beforeEach(()=>{
      spyOn(cpnt.ngUnsub,'next')
      spyOn(cpnt.ngUnsub,'complete')
    })
    
    it(` when called | 
     as appropriate | 
     does the required action `,()=>{
        // act
        cpnt.ngOnDestroy();

        // assert
        expect(cpnt.ngUnsub.next).toHaveBeenCalled();
        expect(cpnt.ngUnsub.complete).toHaveBeenCalled();
    })
  })
});
